Drop support for non-vanity import path
We've supported the vanity import path for this package for quite a
while, and with Go modules, we now have to make the choice between it
and the old import path.

This drops support for the old, non-vanity import path
`github.com/uber-go/atomic` in favor of `go.uber.org/atomic`. Users of
Go modules will have to add the `replace` directive specified in the
README to make this package work with the old import path.