Switch to GitHub Actions from Travis. (#85)
Use GitHub Actions because we're deprecating use of Travis for our OSS
projects.

This change ran into some issues with the `generatenodirty` check so
we changed it to print `git status`. We ran into an issue where
`go mod download` was changing the go.sum and `go mod tidy` was alone.

As a workaround for this, the CI job doesn't `go mod download` by
default; we'll let the actual `test` or `build` commands do that.